trading_bot/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2